Extract EbayParser to own class and share it with popup.js
Popup script initial support for refreshing article info
