+++ utils/build.js
@@ -10,6 +10,6 @@ webpack(
     if (err)
       throw err;
     if (stats.hasErrors())
-      console.log("Errors: stats=%s", JSON.stringify(stats));
+      console.log("Errors: stats=%s", stats);
     // Done processing
   });
